Bug# 40535 : NDBAPI : ScanFilter does not support BIT columns
This bug is fixed to provide ScanFilter support for EQ, NE LT, LE, GT, GE 
on bit columns.

Additionally, 4 new comparison operations are defined for bit types :
 - COL_AND_MASK_EQ_MASK
 - COL_AND_MASK_NE_MASK
 - COL_AND_MASK_EQ_ZERO
 - COL_AND_MASK_NE_ZERO

These can be used to test bits within a bitfield as part of a ScanFiler.

A test program for all of the conditions and bit columns is added