Bug#38793 NDBAPI : EQ never sent for Index scan bounds
NdbRecord ordered index scans currently express an EQ range as separate
lower and upper bounds.  This results in 2 copies of EQ column values
being sent to the kernel which is inefficient.

This patch detects when an equal range is specified by comparing the
passed pointers, key lengths and inclusive bits and sending only one
copy of the equal key columns.  This halves the amount of KeyInfo 
sent for an EQ range.