This commit contains several changes related to queries with a scan operation as root:
* Previously, the API could not know if all results from lookup operations that
 were descendants of a scan had been received. Therefore a change has been made
 such that when the SPJ block sends a SCAN_FRAGCONF message, this will contain 
the total tuple count not just for the scan but also for all descendant lookup 
operations. The API can then compare this count to the number of tuples it has 
received to see if there are more results pending.

* Previously, a read operation that requested the READ_ANY_VALUE pseudocolumn 
would get a correlation number that was guaranteed to be unique among all 
operations derived from a given fragment of the root scan within that batch. 
But this would lead to ambiguities if more than one fragment was scanned in 
parallel. Therefore an extra 32-bit value is returned, to distinguish between 
different fragments in the same batch.

* In the API, a mechanism has been written that links child/parent tuples to 
each other, such that results for child operations are correctly updated when 
iterating over a scan result.

* Results are now handled in the same way for scans and lookups, in the sense 
that the application must call NdbQuery::nextResult() for lookups also. 
(Lookups thus appear to be scans with at most one result row.) 

Observe that:
* RecAttr result retrieval does not work.
* Fetching the next scan batch is not implemented.
* There is no thread synchronization that blocks the application in 
nextResult() yet. The test program therefore does sleep() to wait for results 
to arrive before calling nextResult().