This commit fixes three distinct errors: 1. If a scan operation had an index lookup child, then the tuple count reported to the API (in the SCAN_TABCONF message) would include both index table and base table tuples. Since the index table tuples are never set to the API, the API would hand while waiting for the missing tuples. This has now been fixed. The row count in SCAN_TABCONF will no longer include the index table tuples.
2. If a scan operation had an index lookup child, then the API would 
not be able to match a (base table) tuple from the index operation
with its parent. The base table tuple contained the tuple ID of 
the index table tuple, which again would cointain the tuple ID
of the scan tuple. Since the index table tuple was not sent to
the API. There was thus a "missing link" in the chain of tuple
IDs needed to match the tuples. This has now been fixed by 
making sure that the scan tuple and all its descendants have 
the same tuple ID. (Tuple IDs must only be uniqe within 
operations, not across them.)

3. There was an error in the mechanism for deciding when a scan
operation and all its descendants had finished (a batch). The 
child operations where assumed to be complete when the number
of LQHKEYREQ messages sent was equal to the sum of LHQKEYCONF 
and LQHKEYREF messages received. But sometimes, a TRANSID_AI 
message would arrive after the corresponding LQHKEYCONF and
initiate a new child operation instance. This has no been fixed,
by keeping track of the number of TRANSID_AI messages also.